libkefir: allow inserting rules from end of filter with negative index
If a negative index is passed to the function used to add a rule to a
filter list of rules, start counting from the last element of the list.
This is particularly useful for building a filter object by providing
the rules sequentially, from first to last rule.

+++ src/libkefir_internals.h
@@ -230,6 +230,6 @@ struct kefir_cprog {
  */
 
 int kefir_add_rule_to_filter(kefir_filter *filter, struct kefir_rule *rule,
-			     unsigned int index);
+			     ssize_t index);
 
 #endif /* LIBKEFIR_INTERNALS_H */
